+++ src/Randomizer/EnemyPaletteObject.js
@@ -1,21 +1,14 @@
-/* eslint import/no-webpack-loader-syntax: off */
-import { TableObject, utils } from 'randomtools-js';
-import Color from 'color';
+import { TableObject } from 'randomtools-js';
 import ebutils from './ebutils.js';
 
 class EnemyPaletteObject extends TableObject {
     static shouldRandomize() {
-        return this.context.specs.flags.m; // Enemy stats
+        return this.context.specs.flags.p >= 1; // Palettes
     }
     mutate() {
         super.mutate();
         const hueAdjustment = this.context.random.randint(0, 359);
-        this.oldData.color.forEach((color, i) => {
-            const oldRgb = utils.snesColorToRgb(color);
-            const oldColor = Color(oldRgb);
-            const newRbg = oldColor.rotate(hueAdjustment).rgb().object();
-            this.data.color[i] = utils.rgbToSnesColor(newRbg);
-        });
+        this.data.color = ebutils.rotateEbPalette(this.oldData.color, hueAdjustment);
     }
 }